Move pdeathsig shim to native build
