+++ scripts/build-runtime-helper.sh
@@ -22,6 +22,7 @@ esac
 
 DEST_DIR="$ROOT/packages/native-${PKG_ARCH}-${PKG_OS}/vmm/bin"
 DEST="$DEST_DIR/machinen-runtime-helper"
+PDEATHSIG_DEST="$DEST_DIR/machinen-pdeathsig"
 
 echo "==> building machinen-runtime-helper (zig ReleaseSafe)"
 ( cd "$PKG" && zig build -Doptimize=ReleaseSafe )
@@ -29,11 +30,13 @@ echo "==> building machinen-runtime-helper (zig ReleaseSafe)"
 echo "==> staging into $DEST"
 mkdir -p "$DEST_DIR"
 cp "$PKG/zig-out/bin/machinen-runtime-helper" "$DEST"
+cp "$PKG/zig-out/bin/machinen-pdeathsig" "$PDEATHSIG_DEST"
 
 if [[ "$OS" == "Darwin" ]]; then
   # machinen-runtime-helper does not need entitlements, but clearing provenance
   # keeps the staged host tool consistent with the VMM staging flow.
   xattr -c "$DEST" || true
+  xattr -c "$PDEATHSIG_DEST" || true
 fi
 
 echo "==> Done."
